requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
requests
openai
imapclient
